+++ changelog.md
@@ -1,2 +1,2 @@
 ### Connected Glass 1.1.13
-- Initial release of Connected Glass for Minecraft 1.21.2 & 1.21.3
+- Initial release of Connected Glass for Minecraft 1.21.4